Updated Estimote iOS SDK to version 2.4.0
