+++ plugin/src/ios/Headers/ESTBeaconUpdateConfig.h
@@ -5,7 +5,7 @@
 //  Config object for |ESBulkUpdater|. It describes what needs to be updated
 //  in each beacon. If the value is nil, that beacon's property won't be changed.
 //
-//  Version: 2.3.2
+//  Version: 2.4.0
 //  Created by Marcin Klimek on 06/03/14.
 //  Copyright (c) 2014 Estimote. All rights reserved.
