+++ plugin/src/ios/Headers/ESTBeaconUpdateInfo.h
@@ -2,7 +2,7 @@
 //  ESBeaconUpdateInfo.h
 //  EstimoteSDK
 //
-//  Version: 2.3.2
+//  Version: 2.4.0
 //  Created by Marcin Klimek on 06/03/14.
 //  Copyright (c) 2014 Estimote. All rights reserved.
