+++ plugin/src/ios/Headers/ESTBulkUpdater.h
@@ -2,7 +2,7 @@
 //  ESBulkUpdater.h
 //  EstimoteSDK
 //
-//  Version: 2.3.2
+//  Version: 2.4.0
 //  Created by Marcin Klimek on 16/10/14.
 //  Copyright (c) 2014 Estimote. All rights reserved.
