+++ plugin/src/ios/Headers/ESTDateRule.h
@@ -2,7 +2,7 @@
 //  ESTTimeRule.h
 //  EstimoteSDK
 //
-//  Version: 2.3.2
+//  Version: 2.4.0
 //  Copyright (c) 2013 Estimote. All rights reserved.
 //
